* ping.h    - Be sure we have an 8 bytes integer
git-svn-id: https://tecsw.estec.esa.int/svn/taste/trunk/po-hi-c@4653 129961e7-ef38-4bb5-a8f7-c9a525a55882

+++ examples/aadlv2/ping/ping.h
@@ -1,5 +1,6 @@
 #ifndef __PING_H_
 #define __PING_H_
+#include <stdint.h>
 
-typedef int custom_int;
+typedef int64_t custom_int;
 #endif
